requirements.txt
python-dotenv>=1.0.1
fastapi==0.115.6
uvicorn==0.34.0
python-multipart>=0.0.20
pandas==2.3.3
scikit-learn==1.6.1
joblib==1.5.3
PyPDF2==3.0.1
openpyxl==3.1.5
matplotlib==3.10.8
openai==1.58.1
